Move Definition test to its own file
